Use add_root_certificate
