+++ test/unit/test_sign.py
@@ -20,7 +20,7 @@ import pytest
 from sigstore_protobuf_specs.dev.sigstore.common.v1 import HashAlgorithm
 
 import sigstore.oidc
-from sigstore._internal.timestamping import TimestampAuthorityClient
+from sigstore._internal.timestamp import TimestampAuthorityClient
 from sigstore._internal.trust import ClientTrustConfig
 from sigstore.dsse import StatementBuilder, Subject
 from sigstore.errors import VerificationError
